+++ backend/src/pragma/__init__.py
@@ -10,4 +10,4 @@ Returns:
 Raises:
     None.
 """
-__version__ = "1.0.0b3"
+__version__ = "1.0.0b4"